+++ commercetools/commercetools-sdk-java-api/src/test/java-generated/com/commercetools/api/models/cart_discount/CartDiscountValueAbsoluteTest.java
@@ -26,8 +26,13 @@ public class CartDiscountValueAbsoluteTest {
 
     @DataProvider
     public static Object[][] objectBuilder() {
-        return new Object[][] { new Object[] { CartDiscountValueAbsolute.builder()
-                .money(Collections.singletonList(new com.commercetools.api.models.common.CentPrecisionMoneyImpl())) } };
+        return new Object[][] {
+                new Object[] { CartDiscountValueAbsolute.builder()
+                        .money(Collections
+                                .singletonList(new com.commercetools.api.models.common.CentPrecisionMoneyImpl())) },
+                new Object[] { CartDiscountValueAbsolute.builder()
+                        .applicationMode(com.commercetools.api.models.cart_discount.DiscountApplicationMode
+                                .findEnum("ProportionateDistribution")) } };
     }
 
     @Test
@@ -36,5 +41,15 @@ public class CartDiscountValueAbsoluteTest {
         value.setMoney(Collections.singletonList(new com.commercetools.api.models.common.CentPrecisionMoneyImpl()));
         Assertions.assertThat(value.getMoney())
                 .isEqualTo(Collections.singletonList(new com.commercetools.api.models.common.CentPrecisionMoneyImpl()));
+    }
+
+    @Test
+    public void applicationMode() {
+        CartDiscountValueAbsolute value = CartDiscountValueAbsolute.of();
+        value.setApplicationMode(
+            com.commercetools.api.models.cart_discount.DiscountApplicationMode.findEnum("ProportionateDistribution"));
+        Assertions.assertThat(value.getApplicationMode())
+                .isEqualTo(com.commercetools.api.models.cart_discount.DiscountApplicationMode
+                        .findEnum("ProportionateDistribution"));
     }
 }